Begin extracting attest logic into react hook
