+++ pages/index.tsx
@@ -1,22 +1,15 @@
 import { useAuth, type Auth } from "@/hooks/useAuth";
-import { useMutation } from '@tanstack/react-query';
 import Image from 'next/image'
 import { SiweMessage } from "siwe";
 import { ConnectKitButton } from 'connectkit';
-import { EIP712Proxy } from "@ethereum-attestation-service/eas-sdk/dist/eip712-proxy";
 import type { Address, Chain } from 'viem'
 import { useAccount, useSignMessage } from "wagmi"
 import { Button } from "@/components/ui/button"
-import { jsonParseBigInt } from "@/lib/utils"
 
-import { PROXY_CONTRACT_ADDRESS } from "@/lib/config"
 
 import { isDiamondHands } from "@/lib/diamond-hands"
-import { useSigner } from "@/hooks/useSigner";
-import { useEffect, useState } from "react";
 import { useIsAttested } from "@/hooks/useIsAttested";
-import { ethers, Typed } from 'ethers';
-import { getProxy } from '@/lib/proxy';
+import { useAttest } from "@/hooks/useAttest";
 
 interface SignedOutProps {
   csrfToken: Auth['csrfToken']
@@ -66,52 +59,8 @@ function SignedIn({ session, signOut, csrfToken }: SignedInProps) {
   const githubLinked = session.user?.linkedAccounts?.['github']
   const twitterLinked = session.user?.linkedAccounts?.['twitter']
   const diamondHands = isDiamondHands(session.user?.sub)
-  const signer = useSigner()
-  const [proxy, setProxy] = useState<ethers.Contract| null>(null)
   const isAttested = useIsAttested(session.user?.sub, 'diamond-hand')
-
-  useEffect(() => {
-    if (signer) {
-      setProxy(getProxy(signer));
-    }
-  }, [signer])
-
-
-  const attestMutation = useMutation({
-    mutationFn: async () => {
-      const res = await fetch('/api/attest', {
-        method: 'POST'
-      })
-      const data = await res.json()
-      const response = jsonParseBigInt(data.signedResponse)
-      if (!proxy) {
-        // TODO use toast or something similar to report an error, though I think we should never reach this point
-        return
-      }
-      try {
-        const tx = await proxy.attestByDelegation(
-          Typed.string('diamond-hand'), {
-          schema: response.message.schema,
-          data: {
-            recipient: response.message.recipient,
-            data: response.message.data,
-            revocable: response.message.revocable,
-            expirationTime: 0,
-            refUID: ethers.ZeroHash,
-            value: 0,
-          },
-          attester: response.message.attester,
-          signature: response.signature,
-          deadline: response.message.deadline,
-        })
-        await tx.wait();
-      } catch (err) {
-        // This shouldn't happen but ethers doesn't seem to like
-        // function overloading.
-        console.error(err)
-      }
-    }
-  })
+  const { attest: attestDiamondHands } = useAttest('diamond-hand')
 
   const socialConnections = [{
     name: 'github',
@@ -168,9 +117,7 @@ function SignedIn({ session, signOut, csrfToken }: SignedInProps) {
           {diamondHands ? (
             <><p>You have diamond hands!</p>
               {isAttested ? <p>Already attested</p> :
-              <Button type="button" onClick={() => {
-                attestMutation.mutate()
-              }}>Attest</Button>}</>
+              <Button type="button" onClick={attestDiamondHands}>Attest</Button>}</>
           ) : (
             <p>You do not have diamond hands</p>)}
         </div>
